+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-Development:
+1.5.9:
   - fix issue where wallet mnemonics were not normalised to NFKD
   - "block info" supports fetching the gensis block (--slot=0)
   - "attester inclusion" command finds the inclusion slot for a validator's attestation